app/requirements.txt
altair==4.2.1
av==9.0.1
easydict==1.9
fastapi==0.75.0
nepali_roman==1.1
nltk==3.7
numpy==1.26.3
mediapipe
opencv_python
pandas==1.3.5
Pillow==10.2.0
pytesseract==0.3.9
scikit_learn
scipy==1.12.0
streamlit==1.15.1
streamlit_webrtc==0.45.2
tensorboardX==2.5
tensorflow==2.16.1
torch==2.2.1
torchvision
tqdm==4.63.0
uvicorn
Levenshtein
